Refactoring things out
