+++ src/listeners.js
@@ -1,0 +1,9 @@
+class StreamListener {
+  constructor(producer) {
+    this.next = ev => producer.stream.next(ev);
+    this.error = err => console.error(err);
+    this.complete = () => { };
+  }
+}
+
+module.exports = { StreamListener };
